requirements.txt
psutil>=5.9.0
rich>=13.0.0
customtkinter>=5.2.0
matplotlib>=3.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pystray>=0.19.0
Pillow>=10.0.0
requests>=2.31.0
fpdf2>=2.7.0
